move config utilities to anemoi-training
